requirements.txt
requests
python-telegram-bot
aiohttp
selenium
webdriver-manage